+++ Dimitri/Game/Game/src/Factory/UpdatableFactory.h
@@ -22,7 +22,7 @@ namespace Game {
 			/**
 			* \brief Creates the updatable
 			*/
-			std::shared_ptr<Models::Updatable> create(Enums::TypeEnum type, int x, int y, int z, int height, int width, Enums::DirectionEnum state);
+			std::shared_ptr<Models::Updatable> create(Enums::TypeEnum type, float x, float y, float z, float height, float width, Enums::DirectionEnum state);
 		};
 	}
 }
